FIXED comments from udacity

+++ data_dev/src/mwaa/plugins/operators/load_dimension.py
@@ -20,18 +20,23 @@ class LoadDimensionOperator(BaseOperator):
                  redshift_conn_id="",
                  table_name="",
                  select_statement="",
+                 enableTruncate=True,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id= redshift_conn_id
         self.table_name = table_name
         self.select_statement = select_statement
+        self.enableTruncate = enableTruncate
 
     def execute(self, context):
         self.log.info('LoadDimensionOperator starts')
         self.log.info('getting Redshift Credentials')
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id, autocommit=True)
-        self.log.info('Executing Truncate table: {}'.format(self.table_name))
-        redshift.run(LoadDimensionOperator.truncate_sql.format(table=self.table_name))
+        if (self.enableTruncate):
+            self.log.info('Executing Truncate table: {}'.format(self.table_name))
+            redshift.run(LoadDimensionOperator.truncate_sql.format(table=self.table_name))
+        else:
+            self.log.info('Truncate the dimension table is disabled')
         self.log.info('Executing Bulk insert into {} Dimension Table'.format(self.table_name))
         redshift.run(LoadDimensionOperator.BULK_INSERT_SQL.format(table=self.table_name,select_statement=self.select_statement))
